Started to rewrite heap allocator
This is the beginning of the rewrite of the heap allocator to use the
generic heap data structure we defined earlier today. Right now we have
malloc implemented, meaning we just have to make free. This should be much
easier now that we have a more robust heap data structure