Added more data structures
We now have a linked list, an (almost complete) HashMap, a
barely-started Intervals data structure, and a generic hash API
a la libstdc++.

After we finish the HashMap and Intervals data structures, we can
go on to writing the page allocator. All three of the data structures
we added this commit will be used in that very allocator.