Added Maybe type, Intervals/IntervalSet, pretty printing
Now we have Interval and IntervalSet data structures! The idea is
that we wish to take the memory map from grub and subtract out some
intervals for reserved memory for the kernel/drivers/whatever.

IntervalSet allows us to perform the set operations of subtraction and
union while only having to store the endpoints.

Maybe is just an optional type, but now it, Tuple, Interval, and
IntervalSet all have pretty printing methods with the SerialPrinter