Started work on page directory wrapper
This paging stuff is tricky! So far, we have a functional page
directory wrapper, which we are now populating and using as our new
kernel page directory. This has exposed a number of challenges I did
not quite foresee. Namely: we need contiguous chunk of physical memory
to put our page tables. Hopefully we can figure this out sometime over
the coming week.