Added String, PrintStream classes, updated SerialPrinter
We now have C++-style strings!

Additionally, we refactored most of
the print logic from SerialPrinter to PrintStream. SerialPrinter
allows all of the printing logic from data structures and numeric
types to be reused in any context where one can print a single
character.

Finally, we reworked SerialPrinter to be a bit less cumbersome to
use. Instead of creating a new SerialPrinter object, we can now
just called SP::the()!