Tuple fixes, ELF parsing
As it would turn out, copy constructors are not called automatically
for each field of a class when the class itself has no copy constructor.
This means that when creating a Tuple with a string in it, both strings
would share a common buffer, resulting in a double-free.

Also, we've got ELF parsing! Or at least the beginnings of ELF parsing.
We can now read and verify the sanity of an ELF32 header, prettty print
some of its fields, and also find sections by their name!

Finally, I adapted the stack unwinder from the OSDev wiki, so we
now have a routine to generate very basic stack traces.

The goal of all these components is to finally get line-level stack
traces, meaning the next big goal is to parse enough of the DWARF
sections to be able to translate addreses to lines.